1.Доделал использование анализов в базе данных и сделал на каждый слой свою модель 2. Решил проблему с инициализацией базы данных

+++ app/src/main/java/com/example/wsrpractice/data/storage/dao/PatientCardDao.kt
@@ -11,7 +11,7 @@ interface PatientCardDao {
     suspend fun addCard(patientCardEntity: PatientCardEntity):Long
 
     @Delete(PatientCardEntity::class)
-    suspend fun removeCard(patientCardEntity: PatientCardEntity):Long
+    suspend fun removeCard(patientCardEntity: PatientCardEntity):Int
 
     @Query("SELECT * " +
                 "FROM patient_card_table")
